102 fibonnaci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,17 +8,17 @@
 
 void print_fibonacci(int n)
 {
-	int first = 1;
-	int second = 2;
-	int next;
-	int i;
+	unsigned long long int first = 1;
+	unsigned long long int second = 2;
+	unsigned long long int next;
+	unsigned long long int i;
 
-	printf("%d, %d, ", first, second);
+	printf("%llu, %llu, ", first, second);
 
 	for (i = 3; i <= n; i++)
 	{
 		next = first + second;
-		printf("%d", next);
+		printf("%llu", next);
 
 		if (i != n)
 		{
